# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AcceptSuggestions": {
        "properties": {},
        "title": "AcceptSuggestions",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Linked": {
        "properties": {
          "linked": {
            "title": "Linked",
            "type": "integer"
          }
        },
        "required": [
          "linked"
        ],
        "title": "Linked",
        "type": "object"
      },
...
      "TransferIn": {
        "properties": {
          "inflow_id": {
            "format": "uuid",
            "title": "Inflow Id",
            "type": "string"
          },
          "outflow_id": {
            "format": "uuid",
            "title": "Outflow Id",
            "type": "string"
          }
        },
        "required": [
          "outflow_id",
          "inflow_id"
        ],
        "title": "TransferIn",
        "type": "object"
      },
      "TransferOut": {
        "properties": {
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          }
        },
        "required": [
          "group_id"
        ],
        "title": "TransferOut",
        "type": "object"
      },
      "TransferSide": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "account_name": {
            "title": "Account Name",
            "type": "string"
          },
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "booked_on": {
            "format": "date",
            "title": "Booked On",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "account_name",
          "booked_on",
          "amount",
          "label"
        ],
        "title": "TransferSide",
        "type": "object"
      },
      "TransferSuggestion": {
        "properties": {
          "inflow": {
            "$ref": "#/components/schemas/TransferSide"
          },
          "outflow": {
            "$ref": "#/components/schemas/TransferSide"
          }
        },
        "required": [
          "outflow",
          "inflow"
        ],
        "title": "TransferSuggestion",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kwak Finance API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/transfers": {
      "post": {
        "description": "Mark two transactions as one transfer: no longer spending, income or to categorise.",
        "operationId": "linkTransfer",
        "parameters": [
          {
            "in": "cookie",
            "name": "kwak_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kwak Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Transfer",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/accept-suggestions": {
      "post": {
        "description": "Link every current suggestion.",
        "operationId": "acceptTransferSuggestions",
        "parameters": [
          {
            "in": "cookie",
            "name": "kwak_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kwak Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptSuggestions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Linked"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Transfer Suggestions",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/suggestions": {
      "get": {
        "description": "Pairs that look like a transfer between two of your accounts (same amount, 3 days).",
        "operationId": "transferSuggestions",
        "parameters": [
          {
            "in": "cookie",
            "name": "kwak_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kwak Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransferSuggestion"
                  },
                  "title": "Response Transfersuggestions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer Suggestions",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/transfers/{group_id}": {
      "delete": {
        "operationId": "unlinkTransfer",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "kwak_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kwak Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unlink Transfer",
        "tags": [
          "transfers"
        ]
      }
    }
  }
}